backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
scikit-learn>=1.3.0
PyJWT>=2.8.0
authlib>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
